[SL] Refactor states of code console
