+++ frontend/src/components/CodeRoom/Sidebar/Sidebar.module.css
@@ -25,7 +25,6 @@
   z-index: -1;
 }
 
-
 @keyframes live {
   0% {
     transform: scale(1, 1);
